Move compression code to specific internal files
