+++ libmamba/CMakeLists.txt
@@ -122,6 +122,8 @@ set(LIBMAMBA_SOURCES
     ${LIBMAMBA_SOURCE_DIR}/solv-cpp/queue.cpp
     # C++ wrapping of libcurl
     ${LIBMAMBA_SOURCE_DIR}/core/curl.cpp
+    # C++ wrapping of compression libs (zstd and bzlib)
+    ${LIBMAMBA_SOURCE_DIR}/core/compression.cpp
     # Core API (low-level)
     ${LIBMAMBA_SOURCE_DIR}/core/singletons.cpp
     ${LIBMAMBA_SOURCE_DIR}/core/activation.cpp
@@ -256,6 +258,7 @@ set(LIBMAMBA_PUBLIC_HEADERS
 
 set(LIBMAMBA_PRIVATE_HEADERS
     ${LIBMAMBA_SOURCE_DIR}/solv-cpp/queue.hpp
+    ${LIBMAMBA_SOURCE_DIR}/core/compression.hpp
     ${LIBMAMBA_SOURCE_DIR}/core/curl.hpp
 )
